Moving some functions from model_utils to metrics
