+++ insolver/model_tools/__init__.py
@@ -1,6 +1,6 @@
 from .datasets import download_dataset
 from .metrics import (deviance_poisson, deviance_gamma, deviance_score, deviance_explained,
-                      deviance_explained_poisson, deviance_explained_gamma)
+                      deviance_explained_poisson, deviance_explained_gamma, inforamtion_value_woe, gain_curve,
+                      lift_score, stability_index, lorenz_curve)
 from .model_comparison import ModelMetricsCompare
-from .model_utils import (train_val_test_split, train_test_column_split,
-                          inforamtion_value_woe, gain_curve, lift_score)
+from .model_utils import train_val_test_split, train_test_column_split
